% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantiles.R
\name{qdiff1}
\alias{qdiff1}
\title{Calculate the quantiles of one numeric variable}
\usage{
qdiff1(x, probs)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
Calculate the quantiles of one numeric variable
}
\examples{
#' qdiff1(gapminder$lifeExp)
}

