% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top20GdpPercap.R
\name{Top20GdpPercap}
\alias{Top20GdpPercap}
\title{Select countries with the highest gdpPercap in a particular year}
\usage{
Top20GdpPercap(input_year)
}
\arguments{
\item{input_year}{numeric}
}
\value{
dataframe
}
\description{
Select countries with the highest gdpPercap in a particular year
}
\examples{
#' Top20GdpPercap(2002)
}

